src/mcts_gui/requirements.txt
pywebview~=5.4
py4j~=0.10.9.9